+++ src/domain/usecases/SaveDataSetUseCase.ts
@@ -6,7 +6,6 @@ import { getErrors } from "$/domain/entities/generic/Error";
 import { DataSetUtils } from "$/domain/usecases/common/DataSetUtils";
 import i18n from "$/utils/i18n";
 import { NotificationRepository } from "$/domain/repositories/NotificationRepository";
-import { DataSetRegisterAction } from "$/webapp/components/dataset-wizard/SummaryDataSet";
 import { User } from "$/domain/entities/User";
 import { Config } from "$/domain/entities/Config";
 import { UserGroupRepository } from "$/domain/repositories/UserGroupRepository";
@@ -14,6 +13,7 @@ import { UserGroup } from "$/domain/entities/UserGroup";
 import { ProjectRepository } from "$/domain/repositories/ProjectRepository";
 import { Stats } from "$/domain/entities/Stats";
 import { Project } from "$/domain/entities/Project";
+import { DataSetRegisterAction } from "$/webapp/components/dataset-wizard/DataSetWizard";
 
 export class SaveDataSetUseCase {
     private dataSetUtils: DataSetUtils;
@@ -73,17 +73,16 @@ export class SaveDataSetUseCase {
     private compareOrgUnits(project: Project, dataSet: DataSet): boolean {
         const projectOrgUnits = _(project.orgsUnits)
             .map(orgUnit => orgUnit.id)
-            .sort()
             .value();
 
         const dataSetOrgUnits = _(dataSet.orgUnits)
             .map(orgUnit => orgUnit.id)
-            .sort()
             .value();
 
-        const areEqual = dataSetOrgUnits.every((value, index) => value === projectOrgUnits[index]);
-
-        return areEqual;
+        return (
+            _(projectOrgUnits).difference(dataSetOrgUnits).size === 0 &&
+            _(dataSetOrgUnits).difference(projectOrgUnits).size === 0
+        );
     }
 
     private getUsersGroups(options: SaveDataSetOptions): FutureData<UserGroup[]> {
@@ -132,13 +131,26 @@ export class SaveDataSetUseCase {
     ): FutureData<void> {
         const { dataSet, user } = options;
 
-        return this.getUsersGroups(options).flatMap(userGroups => {
-            const title = `There has been an error when dataset '${dataSet.name}' was being saved.`;
-            const currentUserInfo = `User: ${user.username} (${user.id})`;
-            const body = [title, currentUserInfo, errorMessage].join("\n\n");
+        return this.getUsersGroups(options)
+            .flatMap(userGroups => {
+                const title = i18n.t(
+                    "There has been an error when dataset '{{dataSetName}}' was being saved.",
+                    {
+                        dataSetName: dataSet.name,
+                    }
+                );
+                const currentUserInfo = i18n.t("User: {{username}} ({{userId}})", {
+                    username: user.name,
+                    userId: user.id,
+                    nsSeparator: false,
+                });
+                const body = [title, currentUserInfo, errorMessage].join("\n\n");
 
-            return this.buildUserGroupsAndSendNotification(userGroups, title, body);
-        });
+                return this.buildUserGroupsAndSendNotification(userGroups, title, body);
+            })
+            .flatMap(() => {
+                return Future.error(new Error(errorMessage));
+            });
     }
 
     private getWarningMessages(errorMessage: string): {
