+++ src/webapp/components/dataset-wizard/SummaryDataSet.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { useLocation } from "react-router-dom";
 import { Grid, Typography } from "@material-ui/core";
 
 import { DataSet } from "$/domain/entities/DataSet";
@@ -72,47 +71,3 @@ export const SummaryItem = React.memo((props: { label: string; value: string }) 
         </li>
     );
 });
-
-export const actions = ["edit", "create", "clone"] as const;
-export type DataSetRegisterAction = (typeof actions)[number];
-
-function getActionFromUrl(url: string): DataSetRegisterAction {
-    if (url.includes("edit")) return "edit";
-    if (url.includes("create")) return "create";
-    if (url.includes("clone")) return "clone";
-    throw new Error("Invalid action");
-}
-
-export function useSaveDataSet(props: {
-    dataSet: DataSet;
-    onLoading: () => void;
-    onSuccess: () => void;
-    onError: (error: string) => void;
-}) {
-    const location = useLocation();
-    const action = getActionFromUrl(location.pathname);
-    const { compositionRoot, currentUser } = useAppContext();
-    const { dataSet, onLoading, onSuccess, onError } = props;
-
-    const saveDataSet = React.useCallback(() => {
-        onLoading();
-        return compositionRoot.dataSets.save.execute({ dataSet, user: currentUser, action }).run(
-            () => {
-                onSuccess();
-            },
-            error => {
-                onError(error.message);
-            }
-        );
-    }, [
-        action,
-        currentUser,
-        compositionRoot.dataSets.save,
-        dataSet,
-        onLoading,
-        onSuccess,
-        onError,
-    ]);
-
-    return { saveDataSet };
-}
